Extract Http2StreamBridgeHandler to be extended by HTTP server/client
